Revert peer dependency settings for ESLint
Since ESLint v7 to v8 is potentially breaking, we changed to using `^7` instead of `>=7`. A `^8` version will be created separately. This help to better repos that depend on this package to better manage against regression.

Also took the chance to revert Typescript to `>=4`. Since this is found to be a breaking change, we should increase the major version if this gets accepted.