Revert peer dependency settings for ESLint (#7)
As per the advice of @edward-s.

Since ESLint v7 to v8 is potentially breaking, we changed peer dependency to `^7` instead of `>=7`. A `^8` version will be created separately. This help to better protect against unintended breaking changes in repos that depend on this package.

Also, took the chance to revert-revert Typescript to `>=4` (previously reverted in #6). Since this is found to be a breaking change, we should increase the major version if this PR gets accepted.