Add import sorting
That way we do not need to worry about where an import goes to. It will also automatically merge imports from the same packages.

Side-effect imports are not sorted by default so there is no worry that this sorting will break the code. Refer to the [example in eslint-plugin-simple-import-sort](https://github.com/lydell/eslint-plugin-simple-import-sort#example-1).

Ran the following to install dependencies:
```
npm install --save-dev eslint-plugin-simple-import-sort eslint-plugin-import
```