feat!: revamp ESLint
## context

This package is meant to be a drop in to projects so they can mostly forget about linting. As such, the expected behaviour is for this package to be added, and then the rest of the ESLint dependencies are decided.

The rules have been reviewed and reselected. Notably:
- Stricter rules that can help to prevent bugs are selected.
- Type-checked typescript-eslint rules are used to better flesh out typing issues together with tsc.
- The recommended rule `@typescript-eslint/consistent-type-definitions` is instead customized to prefer type instead of interface. This is aligned with existing practices in the community.
- The `@typescript-eslint/naming-convention` rule is updated to not enforce prepended `I` for interface, but to enforce prepended `T` for generics. Also to align with existing practices in the community.
- Rules are selected for React 16. Make use of the `jsx-runtime` config in later React versions (refer to README).
- Rules are added for Jest files, which should result in much lesser false-positive violations within them.

Projects that disagree can easily turn off anything within their `eslint.config.js`. However, it would be good to make a code change here instead for those that would likely be desirable by more people.

Going ahead ESLint would be using a flat file format, which helps to mix-and-match various configurations, on different sets of files. We applied that in this PR:
1. TypeScript rules are applied to `.ts` and `.tsx` files only.
2. React rules are applied to `.jsx` and `.tsx` files only.
3. Jest rules are applied to JS/TS files within a `test/tests` folder somewhere, or named like a test file, e.g. ending with `test.ts`.

We are unable to upgrade to ESLint 9 yet, as that requires above Node 18.18.0 or Node 20.9.0. Upgrading should be smooth when the time come as we will be using flat config already.

## changes

This is considered a breaking change, so the version is changed to 3.0.0.

Upgraded to ESLint 8. Restructured into flat configs within `index.js`, Removing the older files in the process.

Moved dependencies from `devDependencies` and `peerDependencies` into `dependencies`.

Removed `eslintConfig` from `package.json`. Used `eslint.config.js` instead to construct the ESLint flat config to use.

Added `tsconfig.json` (albeit a sloppy one but it doesn't matter here) as it is necessary for typeScript-eslint to work, since we use the type-checked rules.

Wrote README.md.

## checks

Since this package dogfood the ESLint config within `eslint.config.js`, the following can be verified:

Create a `sandbox.js`. Verify that JavaScript rules are applied to code within.

Create a `sandbox.ts`. Verify that TypeScript rules are applied to code within. Notably, clashing JavaScript rules are turned off.

Create a `sandbox.jsx`. Verify that JavaScript, React, and JSX-a11y rules are applied to code within.

Create a `sandbox.tsx`. Verify that both TypeScript, React, and JSX-a11y rules are applied to code within.

Create test files e.g. `sandbox.test.ts`, `sandbox.spec.ts`, `test/sandbox.ts`, `e2e/tests/random/sandbox.ts`. Verify that Jest rules are applied to code within.